[Finishes #163452432] Delete priveledges given to the admin

+++ app/api/v2/models/models_meetup.py
@@ -49,10 +49,15 @@ class MeetupRecord(BaseModels):
             found = data
         return found
 
-    def delete_item(self, meetupid):
+    def delete_item(self, email, meetupid):
+        user = BaseModels('user_table')
+        user_data = user.find('email', email)
+        isadmin = user_data['isadmin']
+        if not isadmin:
+            return False
         item = self.records.check_exists("meetupid", meetupid)
         if item is None:
-            return False
+            return None
         query = """DELETE FROM meetups WHERE meetupid='%s';""" % (meetupid)
         updated_records = self.records.delete(query)
         return updated_records
